+++ PluginsAndFeatures/azure-toolkit-for-rider/CHANGELOG.md
@@ -4,23 +4,31 @@
 
 ## [Unreleased]
 
+### Added
+
+- WebApp and Function property views ([#767](https://github.com/JetBrains/azure-tools-for-intellij/issues/767))
+
 ## [4.0.0-preview.3] - 2024-03-22
 
 ### Changed
+
 - Support for Rider 2024.1
 
 ## [4.0.0-preview.2] - 2024-02-05
 
 ### Changed
+
 - Update tool window icon
 - Reimplement Azurite support
 
 ### Fixed
+
 - Unable to deploy to the existing WebApp ([#782](https://github.com/JetBrains/azure-tools-for-intellij/issues/782))
 
 ## [4.0.0-preview.1] - 2024-01-24
 
 ### Changed
+
 - Reimplement Azure account functionality
 - Reimplement Azure Explorer tool window
 - Reimplement Azure Web Apps and Azure Web Apps for Containers deployment
@@ -30,6 +38,9 @@
 - Reimplement Azure Functions templates
 
 [Unreleased]: https://github.com/JetBrains/azure-tools-for-intellij/compare/v4.0.0-preview.3...HEAD
+
 [4.0.0-preview.3]: https://github.com/JetBrains/azure-tools-for-intellij/compare/v4.0.0-preview.2...v4.0.0-preview.3
+
 [4.0.0-preview.2]: https://github.com/JetBrains/azure-tools-for-intellij/compare/v4.0.0-preview.1...v4.0.0-preview.2
+
 [4.0.0-preview.1]: https://github.com/JetBrains/azure-tools-for-intellij/commits/v4.0.0-preview.1